+++ docs/js/worlds.js
@@ -10,25 +10,28 @@ async function getWorlds(callback) {
   });
   worldJson = await response.json();
 }  
+function formatOutputTime(input1, input2, arrayPrep) {
+  if (input1 < 10 && input2 < 10) {
+    return ("0" + String(input1) + ":0" + String(input2));
+  }
+  else if (input2 < 10) {
+    return (String(input1) + ":0" + String(input2));
+  }
+  else if (input1 < 10){
+    return ("0" + String(input1) + ":" + String(input2));
+  }
+  else {
+    return (String(input1) + ":" + String(input2));
+  }
+}
 async function makeArray() {
   for (i in worldJson['servers']) {
     let dateDiff = parseInt((Date.now() - worldJson['servers'][i]['firstSeen'])/1000);
     let arrayPrep = [];
     arrayPrep.push(String(i));
-    if (Math.floor(dateDiff%3600/60) < 10 && Math.floor(dateDiff/3600) < 10) {
-      arrayPrep.push("0" + String(Math.floor(dateDiff/3600)) + ":0" + String(Math.floor(dateDiff%3600/60)));
-    }
-    else if (Math.floor(dateDiff%3600/60) < 10) {
-      arrayPrep.push(String(Math.floor(dateDiff/3600)) + ":0" + String(Math.floor(dateDiff%3600/60)));
-    }
-    else if (Math.floor(dateDiff/3600) < 10){
-      arrayPrep.push("0" + String(Math.floor(dateDiff/3600)) + ":" + String(Math.floor(dateDiff%3600/60)));
-    }
-    else {
-      arrayPrep.push(String(Math.floor(dateDiff/3600)) + ":" + String(Math.floor(dateDiff%3600/60)));
-    }
+    arrayPrep.push(formatOutputTime(Math.floor(dateDiff/3600),Math.floor(dateDiff%3600/60)));
     arrayPrep.push(String(Object.keys(worldJson['servers'][i]['players']).length));
-    arrayPrep.push(String(20 - Math.floor(dateDiff%3600/60)%20));
+    arrayPrep.push(20 - (Math.floor(dateDiff%3600/60))%20);
     finalArray.push(arrayPrep);
   }
 }