+++ docs/changelog.md
@@ -5,6 +5,7 @@ permalink: /changelog/
 ---
 ## 2021/11/3
 - World list [soul point regen default sort](/sptime/)
+- cleaned up worlds.js
 
 ## 2021/10/27
 - Quest status update